+++ client/routes.js
@@ -1,0 +1,41 @@
+
+import React, { Component, PropTypes } from 'react';
+
+import { Router, Route, hashHistory, IndexRedirect } from 'react-router';
+import { syncHistoryWithStore } from 'react-router-redux';
+
+import AppContainer from './containers/App';
+import StatusPage from './containers/StatusPage';
+import DebugPage from './containers/DebugPage';
+import AccountsPage from './containers/AccountsPage';
+import AppListPage from './containers/AppListPage';
+import RpcPage from './containers/RpcPage';
+import RpcCalls from './components/RpcCalls';
+import RpcDocs from './components/RpcDocs';
+
+export default class Routes extends Component {
+
+  render () {
+    const history = syncHistoryWithStore(hashHistory, this.props.store);
+    return (
+      <Router history={history}>
+        <Route path={'/'} component={AppContainer}>
+          <IndexRedirect to='status' />
+          <Route path={'status'} component={StatusPage} />
+          <Route path={'debug'} component={DebugPage} />
+          <Route path={'accounts'} component={AccountsPage} />
+          <Route path={'apps'} component={AppListPage} />
+          <Route path={'rpc'} component={RpcPage}>
+            <IndexRedirect to='calls' />
+            <Route path={'calls'} component={RpcCalls} />
+            <Route path={'docs'} component={RpcDocs} />
+          </Route>
+        </Route>
+      </Router>
+    );
+  }
+}
+
+Routes.propTypes = {
+  store: PropTypes.object.isRequired
+};
